back-end/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bcrypt==4.0.1
black==25.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.3.0
cloudinary==1.44.1
colorama==0.4.6
cryptography==46.0.3
dataclasses-json==0.6.7
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.119.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
filelock==3.20.0
flake8==7.3.0
frozenlist==1.8.0
google-ai-generativelanguage==0.4.0
google-api-core==2.26.0
google-auth==2.41.1
google-generativeai==0.3.2
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
hiredis==3.3.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
identify==2.6.15
idna==3.11
isort==7.0.0
Jinja2==3.1.6
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.2
langchain-classic==1.0.0
langchain-community==0.4
langchain-core==1.0.0
langchain-google-genai==0.0.1
langchain-text-splitters==1.0.0
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-checkpoint-postgres==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.37
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mccabe==0.7.0
mdurl==0.1.2
multidict==6.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==2.3.4
openai==2.6.0
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pgvector==0.4.1
platformdirs==4.5.0
pre_commit==4.3.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.25.8
psycopg==3.2.11
psycopg-binary==3.2.11
psycopg-pool==3.2.6
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyflakes==3.4.0
Pygments==2.19.2
PyMySQL==1.1.2
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.2.0
PyYAML==6.0.3
redis==7.0.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
rsa==4.9.1
sentry-sdk==2.42.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
tenacity==9.1.2
tqdm==4.67.1
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
# uvloop==0.22.1
virtualenv==20.35.3
watchfiles==1.1.1
websockets==15.0.1
wheel==0.45.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
